Decrease max packet size for sending
The max packet size is reduced to 4096, to more closely approximate the
MeerKAT CBF. The CBF uses a max *payload* size of 4096, but spead2
doesn't have an option to limit payload size independently from packet
size.

This will probably be superseded by #4 (which brings a command-line
option to configure it) once it is ready, but this change is made
for now to reduce the packet size below the new SDP MTU of 9000.