First work on adding polarization support
A new Source class is added that allows for adding polarization
information in addition to existing katpoint target properties. It is
generic to allow other properties to be added later (e.g. to support
non-point sources).

The Stokes parameters are converted to a linear-polarization brightness
matrix. However, the parallactic angle rotation is not taken into
account.