Overhaul visibility simulation to handle feed rotation
This overhauls the visibility simulation to move all Jones matrices into
the predict step. Previously gains were handled after sampling, but for
general Jones matrices that won't lead to the correct result.

_flux_sum was eliminated, because the computation used for it wasn't
correct for general Jones matrices. The correct value can be read out
of the visibility matrix, which is what is now done.

Feed rotation is simulated, including a correction for (V, H) in MeerKAT
winding the opposite way to celestial (x, y).