Fix substream count when antennas added after destination set
Previously the number of substreams was set at the time the capture
destination was set. In the katsdpcontroller-driven workflow, that
occurred before configuring antennas, causing the wrong number of
substreams to be used.

Determining the number of substreams is now deferred to the time the
transport is constructed. For implementation reasons, that also means
that specifying more endpoints than the default number of substreams now
requires the user to explicitly increase the substream count.