Refactor how antennas are set
The change to set telstate during stream creation, rather than on
capture-init, meant that the number of antennas must be fixed when the
stream is created, even though the descriptions might be not be
available from kattelmod yet.

To facilitate this, a number of changes are made
- katsdpcontroller sets antenna_mask in config, and this is now used to
  fake antennas with the correct names but made-up positions, just to
  get the count correct.
- Subarray now prevents certain attributes being changed once a stream
  has been created on it. Previously this was allowed as long as no
  capture was in progress, but wouldn't have the desired effect because
  the values are immutable in telstate.
- configure_subarray_from_telstate is now stricter, and will raise an
  exception if the telstate keys are missing, rather than just warning.

I also did a bit of cleanup on the exceptions, creating a base
StreamError exception, generalising IncompleteConfigError to
ConfigError, and using ConfigError in some cases that previously used
ValueError.