Add delay to synchronise multiple cbfsim servers
See the comments in the code for how this works. It's a little hacky and
it may be better to synchronise with some extra primitives in telstate,
but that probably needs to wait until we have programme block IDs that
can be useful to form a key that is shared between all the instances of
the simulator per capture-start.