Use descriptors to simplify adding new subarray properties
The Subarray class had a lot of boilerplate code to restrict when
certain properties could be set. This provides a descriptor class to
automate that, which will allow new properties to be added with less
boilerplate. It doesn't add any new features, although it fixes the
stream-immutable properties to raise StreamExistsError as documented
instead of CaptureInProgressError.