Provide control over the stream start time
Previously the stream was always simulated from the sync time - the
first dump had an ADC timestamp of 0. That is problematic for testing
with kattelmod because it meant that successive program blocks all
simulated from the same point, which confused things like timeplot that
saw time go backwards. It also introduced problems because of the
discrepancy between the time being simulated in kattelmod and that being
simulated in katcbfsim.

The start time is now controlled by the new argument to ?capture-start:
if not given, you get the old behaviour for backwards compatibility.
Otherwise, you specify a timestamp for start (or "now").

For completeness, cbfsim.py now takes a --cbf-start-time for cases where
--start is given, but the intention is that master controller will pass
time.time() to capture-start.